Lift `mock` module out of `mesh`
It is large enough to have its own file.
